Ref ticket:74 Disable mutex locking until progress bar can be rewritten.
The problem here seems to be how Qt is queueing the signals. The
assertion claims that the mutex is being locked and unlocked from
different threads, but the QueuedConnection type should prevent that
from happening, and inspecting the current thread IDs on each
lock/unlock reveals that they are in fact the same each time.

What is happening? One thread requests to start a progress update,
which is handled by the GUI thread. Before this progress finishes,
another thread requests a progress update, which finishes before the
first update finishes. The dialog then handles both unlock calls
before handling the second thread's lock call, essentially unlocking
an unlocked mutex.

The framework for the progress bar will need to be rewritten. For now,
just remove the lock/unlock calls. This may result in bizarre behavior
if simultaneous progress updates are requested, but nothing more than
a GUI glitch, not a crash.