Made a note about a resume bug.
From XtalOpt::load:

TODO For some reason, the local view of "this" (e.g. this->tol_spg) is
not changed when the settings are loaded in the following line. The
tabs are loading the settings and setting the variables in their
scope, but it isn't changing it here. Caching issue maybe?