Don't timeout remote execution requests. If the command produces no output, a false positive occurs.
The timeout period has been shortened to 500 ms and a timeout is no
longer fatal in SSHConnection::_execute. This is to give a reponse a
better chance of being received while allowing for there to be no
reply.

The exit code check will still timeout, since this is an expected
reply.