Two modifications to SSHConnection to help deal with large outputs.
First, there was a bug in the channel_read commands where
count=bytesAvail, even though it was possible that bytesAvail could
exceed the buffer size. This is corrected.

Secondly, the TCP/IP stack could fill and, once emptied, new data may
not appear in time for channel_poll to detect it. A one second timeout
is introduced to the while(channel_read(...)){...} loop to remedy this.