Reverted recent changes in SSHConnection::execute.
The new implementation was fickle about timeout and extremely slow if
there was no stdout / stderr. The code is this commit is more than an
order of magnitude faster on my benchmarks, and is not susceptible to
latency vs. timeout races.

So long as the ssh shell + timeout approach is used in
SSHConnection::isConnected, VPN timeouts should still be detected
appropriately.