Added a 25x max angle different requirement
If the angles of a generated xtal are very small, then, after Niggli reduction,
one of the sides may be adjusted to be very small instead. This can cause
spglib to crash. It's reasonable to not expect one side to be at least 25x greater
than another side. So let's enforce that requirement.