Create valid lattice before adding atoms in
Now, a valid lattice is created before atoms are added
into it. This can speed up xtal generations. It will also
prevent molUnits from rescaling so that their suggested
length is correct.

Unfortunately, molUnit distances still get rescaled when
we use mitosis. This should perhaps be fixed in the future.