Removed the final dependence on OpenBabel
XtalOpt is now fre from dependence on OpenBabel and
Avogadro! However, not everything is complete. The plot
tab isn't working and needs to be constructed using new
plotting software. MolUnit does not work and needs to be
fixed using new functions to create the molUnits. We no
longer have any formats: we need to re-write the formats
to be read. And finally, there are probably bugs that need
to be fixed. A lot of testing needs to be done.