Added VASP format - it seems to work properly
Further tests can be done. But from an initial glance, it looks
like it is working fine. I also fixed some bugs with seeding
structures.

There is a bug: if the user presses "begin" several times,
XtalOpt will get confused and will not run. We should make
it so that they cannot do that.