Re-added plotting with Qwt
We now have a plot tab again, and plots are
generated using Qwt. This addes Qwt as a dependency,
but it is usually easy to obtain. Most features from
the previous plot are present including clicking on points
to highlight them (and select them on the progress tab),
labelling points, plotting trends, etc.

The distance histogram does not work and will be removed soon.