Updated to work on Ubuntu 14.04
CMake now runs properly with earlier versions of
cmake.

There is one problem, however: qwt-qt5-dev is not
available on earlier OS's like Ubuntu 12.04 and
Ubuntu 14.04. So, either the user has to compile
qwt by source, or we will provide them with some
precompiled stuff. We will probably do the latter.

If the user tries to use the wrong qwt, xtalopt will
just segmentation fault when they try to run it (there are
no compilation errors).