Fixed a crashing bug in AbstractEditTab
There appears to be some sort of race condition in
abstractedittab. Sometimes, the function 'updateEditWidget()'
crashes when ui_list_optStep->currentRow() is -1. It should
definitely not be -1, but I am unsure as to why it is correct
in some runs and incorrect in others. I set up an if statement
that guards against crashing the program. Hopefully, it doesn't
cause any other issues. But returning to avoid crashing appears
to work just fine (even though the function is not re-called).