Fixed multithreading unnecessary debug message
The message that was fixed is as follows:
QObject: Cannot create children for a parent that is in a different thread.
(Parent is GlobalSearch::LocalQueueInterface(0x24c3540), parent's thread is QThread(0x2147350), current thread is QThread(0x21edcb0)

This fix will hopefully not introduce other bugs.

In addition, this only fixes the problem for the
'local' queue interface. I'm not sure if the bug
is present in other queue interfaces... but it
will not be fixed in other queue interfaces if
it exists there.