Fixed queue manager list not updating bug
This fixes a race condition bug that caused
first few xtalopt jobs to not start properly.
I am not sure why QueueManager was only moved
into its separate thread once the thread started,
but it no longer does that. Now, QueueManager is
immediately moved to its separate thread upon
construction.

Hopefully this does not introduce any bugs. But it
does not seem to...