Updated random number generators
A new, more straightforward random number
generator has been implemented and replaced the
older one. <globalsearch/random.h> needs to be included
to use it.

Now, getRandDouble() is replacing RANDDOUBLE().
getRandDouble() generates a random double between 0.0 and 1.0.
However, you may put in extra parameters to specify a range.
For instance, getRandDouble(-1.0,5.2) will generate a random
double in that range.

getRandInt() is similar and you may specify a range. And
getRandUInt() is similar as well.

No INIT_RANDOM_GENERATOR() is necessary.