Added .travis.yml for Linux travis-CI builds
There is now a travis-CI build on Ubuntu 14.04.
Unfortunately, travis-CI does not yet have Ubuntu 16.04.
To get around issues, I'd either need to use docker and
a virtual Ubuntu 16.04 environment, or I'd need to use
my .deb file for installing qwt-qt5-6 on Ubuntu 14.04.

I chose the .deb file for installing qwt-qt5-6 on Ubuntu 14.04.
Thus, it uses wget to obtain the .deb file, and then it installs
qwt-qt5-6 with it. Seems to work fine.