Plot now only updates at the end of a load
Added signals and slots so that the plot tab is not
updated while resuming a job. This saves a lot of
time.

Also removed the unneeded populateXtalList()
function since it was only used for the histogram.
We can bring it back if we decide we want to use
histograms again. But removing this function also
saves time on jobs.

Also removed the std::atomic_bool from xtalopt
and started just using the boolean already present.

Some more work needs to be done to improve speed of
resuming jobs, but this is not bad.