Fixed a bug with loading opt schemes
Fixed a bug that occurred when the POTCAR template was
the first template in the m_templates hash of Optimizer. When
the POTCAR template was the first, the number of opt steps would
come out to be 1 instead of the correct number. Now, it finds the
largest template and counts that as the number of opt steps.

There is still a bug when adding templates. For some reason, it
adds two instead of one and seg faults if you click on the second
one...