Added 'make install' for XtalOpt on Linux and MSVC
A CMake script was created that searches for all dependencies
and optionally installs them. In addition, the changelog, readme,
licensing, and sample schemes are installed. 'make package' now
runs cpack, but it doesn't really do anything useful. We will need
to soon get cpack to also install the dependencies and set up
all configurations for it.