Automatically re-run cmake if INSTALL_DEPENDENCIES
Automatically re-run cmake if INSTALL_DEPENDENCIES is set
and xtalopt has not yet been built. Then, the dependencies
can be correctly found and installed.

The way INSTALL_DEPENDENCIES is set up requires there to
be an XtalOpt executable before it can locate the
dependencies since it finds the dependencies via the
executable.