Fixed NSIS packaging and added AutoRCC and AutoUIC
Now the NSIS packaging has the XtalOpt logo on it
correctly and it allows for a shortcut to be placed on
the desktop.

In addition, the MSVC executables now correctly have the
XtalOpt logo.

In addition, I deleted all the functions and variables
involving qrc files and ui files and enabled AutoRCC and
AutoUIC to take care of all of it.