Re-wrote generateSuperCell() and cleaned up
This is in an attempt to help fix some deadlocks
we have been facing. generateSuperCell() appears to
be working fine, and I haven't experienced a deadlock
yet with this setup.

The QReadWriteLocks were changed to be recursive to
improve their stability (so we can perform a lockForRead()
multiple times in the same thread without accidentally
deadlocking if another thread calls a lockForWrite()).

Also, a lot of lockers were used.