no-gui feature added
It seems to be working fine locally with GULP. There
are, however, some crashing issues with VASP.

A lot more testing needs to be done.

You run the no-gui version by adding the command
line options "--no-gui" after the executable. You can
also specify the input file with "--input-file=". A sample
input file is now given in samples: xtalopt-cli.in. That is
the default name of an input file.