Created molecule.cpp and fixed bond decrement bug
Several functions in molecule.h had become larger than inline
functions should be, so a molecule.cpp file was created, and
those functions were placed in there.

In addition, a few lines were added to Molecule::removeAtom(size_t ind)
so that all atom indices in the bonds that are greater than the atom
index that was removed will be properly decremented (so that they
still point to the correct atom).