Added bond perceiving function
The bond perceiving function will generate a bond between two
atoms if the distance between them is less than the sum of their
covalent radii plus a tolerance factor.

This was tested on several small molecules and appears to work fine:
acetylene, benzene, ethylene, methane, and water. Without the
tolerance factor of 0.1, no bonds in water were detected.

The tolerance factor may be changed in the future.