Added a basic RandSpg test
Right now, it just makes sure that RandSpg can generate Xtals.
We should add more tests with this in the future.

Strangely enough, it seems that spglib doesn't detect the correct
space group almost 50% of the time. I think this is a spglib problem,
though, because RandSpg was tested thoroughly with the golden
standard: findsym.

If we figure out why spglib isn't detecting the space group correctly
in the future, we can add in a spglib confirmation.