Moved POSCAR readers/writers to formats dir
Removed the Xtal functions that read from a
POSCAR and return an Xtal*, and move all POSCAR
reading functions to a specific function in the
formats directory.

Note: this POSCAR reader MUST read the atomic
numbers, or else it will fail.

I also went through the tests and removed a
few pointers to Xtals that were being
used in conjunction with the POSCAR readers.

I also added tests for the POSCAR reader and
writer.