Major: moved reading/writing settings to XtalOpt
Now most settings (except for GUI-exclusive ones) are
read and written from XtalOpt itself. This allows XtalOpt
to function without the GUI dialogs. In fact, a
QCoreApplication is now created instead of a QApplication
when CLI mode is being ran.

This is not completely bug-free. Some new bugs have to be
worked out. But it appears to be working for the most part.