Add molecules randomly to unit cell and tests
Added function that adds molecules randomly to a unit cell
(including checking IADs - both intramolecular and intermolecular).

Also added a basic minIADs class that uses unordered pairing to
map a pair of atomic numbers with a distance.

Also added a function to set the mean position of a molecule.

The test seems to work fine. But more testing will be needed in
the future.