Major change: different opts and QIs for steps
Updated the program so that a different optimizer and
a different queue interface may be used for every step.

This is a major change that renders previous scheme and state
files incompatible with the current version.

I updated the sample schemes for xtalopt to reflect these changes.
They can be used for new schemes.

In addition, the old method for selecting POTCAR files is no longer
valid. Instead, the user needs to put in keywords as such:
"%filecontents: <path_to_file>%". It will replace the section
delimited with the percent signs with the contents of the file. So
the user can put down multiple lines of it to use multiple POTCAR files.
Note also that the elements need to match the order that they are given
on the init tab. Same is true for siesta potential files.

CLI mode with these changes hasn't been tested. But the GUI mode seems
to work.