Use SD format instead of PDB for conformers
Unfortunately, PDB cannot describe bond orders. When RDKit
performs MMFF94 optimization, it needs correct bond orders
to perform optimization correctly. So we needed to change
the format used for conformers, and we are now using the
SD format.

In addition, we started using RDKit's getBestRMS() and
deleted the one I had written.

Also added a butane.mol file to the samples.