Got the Siesta z-matrix writer working
I tested this a few times with SIESTA and it seems
to be working fine.

It was tricky because SIESTA expects some extra information
to relay position and orientation.

The first atom needs to give its Cartesian coordinates.

The second atom needs to give its spherical coordinates (with the
first atom being the origin).

The third atom uses an imaginary atom to calculate the dihedral.

In my comparisons, this commit seems to be working fine.