Added a "copyfile" keyword to templates
Now, a user may use the "copyfile" keyword in a template
as follows:

%copyfile:/path/to/file%

If the user does this, that file will be copied into the
working directory for each crystal when the templates are
created.

We can immediately use this for .psf files in siesta. The
user can put in %copyfile:/path/to/psf/file% for a psf file
for each element. This will copy the psf files to the directory
that siesta will be ran in.