Added pre-optimization bonding info storage
Added the ability to store pre-optimization bonding information
and re-use it after optimization.

This is very useful for optimizers that do not store bonding
information. As long as the ordering of the atoms stays the same,
we can keep the same bonding information after optimization.

This is currently implemented for the Siesta z-matrix, but nowhere
else.