Fixed a few bugs in siesta spherical coords writer
Sometimes, the molecule would be oriented in the wrong direction
due to a bug in the spherical coordinate writer for the siesta
z-matrix. This fixes the issue.

However, the spherical coordinate writer does not yet appear to be
completely bug-free. There is a rare bug where a hydrogen seems to
be misplaced. Need to look into it more.