Fixed more siesta z-matrix calculation bugs
Molecule::angle() and Molecule::dihedral() no longer
use the minimum images because they led to bugs.
Minimum images can be valuable, but when used between
atoms that are not bonded together (like some atoms in
dihedrals), it gives you incorrect results.

To fix this, the siesta z-matrix calculator now wraps
all atoms to their shortest bonds, and then it performs
all of the calculations as normal.

This was tested on about 40 butane molecules, and each
butane came up in the correct orientation and position.
It appears that the siesta z-matix calculator works
correctly now.