Added ability to calculate hardness from xtals
There are two checkboxes that have been added to the
"Search Settings" tab: "Calculate hardness with Aflow ML?"
and "Use hardness for fitness function?"

"Use hardness for fitness function?" doesn't do much right now.
But if "Calculate hardness with Aflow ML?" is checked, when a
structure emits "finished", it will be submitted to the Aflow
servers to have its bulk and shear moduli calculated.

This is not bug-free, however. "Calculate hardness with Aflow ML?" has
been show to cause deadlocks for currently unknown reasons. And
the plot for some reason plots the vickers hardnesses as 0 if they are
not present. More bug research needs to be done.