In Molecule, changed reorderAtoms to sortAtoms...
and added reorderAtoms. sortAtoms() must have a
number for each atom from 0 to atoms.size() - 1, as must
reorderAtoms(). The difference is that sortAtoms() will sort
the atoms so that the numbers are 0 - atoms.size() - 1, but
reorderAtoms() will replace each atom with the atom that has
the specified index in the list.

Also fixed siesta z-matrix reordering to reflect these changes.