Add format functions for reordering atoms
For z-matrix and poscar, added functions to reorder atoms to match
that of the z-matrix and poscar formats. This enables us to preserve
bonding information during the optimization period. Also added
functions to perform this reordering if the variables are set to do
so.

This is untested currently. It needs to be tested to show that it
works.