Use an atomic size_t for Aflow::m_requestCounter
This removes the need to lock the mutex when submitting
a request (to hopefully prevent GUI hangups).

Also used a std::unique_ptr for AflowML object so that its
header file can be placed in the .cpp file (to prevent a
need to recompile everything when we change aflowml.h).