Added hardness fitness function
Made some modifications to OptBase::getProbabilityList()
so that a probability list based on hardness may be generated
instead. This equation is similar to the enthalpy probability
equation except the numerator of the fraction is:

(H_max - H_i)

instead of

(H_i - H_min)

This effectively results in a probability list where there is
a higher probability fo picking a value with a higher hardness.

If you pass in "useHardness = true", the list must be sorted by
hardness instead of enthalpy.

These functions are not necessarily written in the best way,
but I have not made the effort to go through them and clean them
up.