Added dialog for XRD plotting, doesn't fully work
A context menu item has been added to the progress tab. When
an xtal is right-clicked, an option to plot a theoretical XRD
pattern is displayed. If clicked, a dialog box appears where the
user can set options (that will be saved for the current run but
will always be reset between runs).

When the user clicks 'ok', a theoretical XRD pattern is calculated,
but it is not yet plotted. It will be plotted in the next commit
involving the XRD pattern plotting.