Added another QwtPlot to plot XRD Patterns
The XRD patterns can now be successfully plotted by
right clicking an entry in the progress tab and selecting
to plot the XRD pattern. It seems to be working fine on Linux.

There are some issues on Windows and Mac, though, as can be seen
by the failing Travis-CI and failing AppVeyor. So those will need
to be resolved soon.