Resubmit unfinished hardness calcs every 10 mins
When "calculate hardness?" is checked, another thread will
now be started that resubmits unfinished hardness calculations
every 10 minutes. This is to go back and clean up unfinished hardness
calculations.

This is the first case where we used std::thread from the standard
library instead of Qt's threads. If we run into linking problems
or packaging problems, we can always go back to QtConcurrent.