Progress tab: context menu "replace with..." fixed
Since we renumbered the opt steps to start from 0,
the "replace with..." context menu options in the progress
tab have caused the program to crash if there is only one
opt step. This is because it tries to restart the crystal
on opt step 1, which does not exist if we only have one
opt step.

This update fixes the problem. "Replace with..." now appears
to work properly.