Added hardness weight to the fitness function.
If a user is calculating hardness, they may use a
weight between 0 and 1 to specify the weight of the hardness
on the probability selection.

If a user selects 0, the probability selection will be based
entirely upon enthalpy (where low enthalpy is favored).

If a user selects 1, the probability selection will be based
entirely upon hardness (where high hardness is favored).

Anything between 0 and 1 will be based on a combination of enthalpy
and hardness where low enthalpy and high hardness are favored. The
closer the number is to 0, the more important low enthalpy is (over
high hardness), and the closer the number is to 1, the more important
high hardness is (over low enthalpy).

The equation for the fitness function is now as follows:

pi = N [ 1 - w (Hmax - Hi) / (Hmax - Hmin) -
             (1 - w) (Ei - Emin) / (Emax - Emin) ]

where H is the hardness, E is the enthalpy per formula unit,
w is the fractional weight of the hardness (between 0.0 and 1.0),
and N is a normalization constant.