Improved logic and speed of duplicate matching
In order to fix issues we have been facing with the
gene pool getting flooded with the same crystal (such as
diamond), we needed to change the logic of some of the duplicate
matching. I also improved the efficiency of the duplicate matching
while I was working on it.

The bug was that: if a crystal was found to be primitive-reducible,
a primitive reduction would be performed and a new crystal generated.
The first crystal would always be set as a supercell of the second
crystal. This means that if several 8-FU diamond crystals were found,
they would all be a supercell of their own 2-FU primitive-reduced
diamond crystals. This would result in each of those 8-FU diamond
crystals being placed in the gene pool (because each one would be
a supercell of a different crystal).

The logic has been changed. Now, if a crystal is identified as a
duplicate, it will not be identified as a supercell. Indeed, the
only crystals that will certainly be identified as a supercell of
their primitive-reduced forms are the ones that were not found
to be a supercell automatically.

Supercells will now more frequently be identified as duplicates.
This seems to have fixed the gene pool issue (it has been tested).

In addition, the efficiency of the duplicate-checking algorithm
seems to be improved.