Queue: Generalized Global Queue Interface Settings
I re-worked the global queue interface settings so that it has its
own widget and is re-used in all of the queue interfaces. This reduces
a lot of repeated lines of code and improves re-usability.

In addition, I added a new option in the global queue interface
settings that allows the user to cause a job to be killed if it
exceeds a user-specified amount of time.

This "time" is both queued + running time.