Bug Fix: Load settings from crashed xtalopt run
Previously, resuming a crashed XtalOpt run wouldn't load all of
the settings properly. This is because signals and slots would
get fired for updating the GUI during a resume, and the settings
would get over-written.

This fixes the problem for the most part, but more testing is needed.
It may require a little more work.