Fixed the cli mode as it was broken
The cli mode now works (at least for GULP) as before. Multiple
optimization steps are allowed. However, each optimization step
must use the same optimizer and queue interface. We should eventually
update this in the future so that we can use different ones for each
step (as you can in the GUI).

Testing of VASP still needs to be performed.