Fix application starting issues on Apple
It seems that if we run "open" on the Apple application bundle, it
gives a cryptic error message and fails if the QApplication is not
instantiated before running the QCommandLineParser options are set.

Strangely enough, the executable by itself seems to run just fine.
We only run into an issue when we double-click on the application
or start it via the "open" command.

To fix this, we are now reading some command line options manually
to determine whether we are in CLI mode or not, and then instantiating
the QCoreApplication or QApplication after that. Then we can read
the rest of the QCommandLineParser options.