Prevent infinite loop when changing potcar in run
This bug was observed on Ubuntu 14.04 - I haven't seen it elsewhere.

The bug was: changing the text of the POTCAR in the middle of a run
would cause a crash.

Essentially, saveCurrentTemplate() would perform a font change, and
the font change would trigger a textChanged() signal which would
re-fire saveCurrentTemplate(). This fix should prevent that issue.