Add better error detection to generic optimizer
Now, the generic optimizer will check that the following are true
in the updated structure (after optimization is complete):

1. The structure has more than zero atoms.
2. The structure has a unit cell.
3. The structure has energy or enthalpy.

If any one of these is false, then the generic optimizer treats the
optimization as a failure.

Still, not every error will be caught, because sometimes
Open Babel will read all the parts successfully from an output
file even though the job failed. In this case, the XtalOpt user
can write lines in the job script after the optimizer step to
make sure that the output is complete - and just rename the output
to something other than "job.out" if the output is not complete.