Use correct base URL for AFLOW-ML
The endpoint "aflow-ml.materials.duke.edu" was used previously,
which is tied to a specific server. The virtual machine was moved
to a different server, however.

Use "aflow.org/API/aflow-ml/v1.0" instead as it is directed to the
correct server.

This fixes the issues where the http requests time out when trying
to calculate the hardness.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>