Attempt to stop infinite log bug
This is an attempt to stop a bug that is occurring very rarely, but when
it does occur, it prints to a log file non-stop until the disk runs out
of memory. This is very bad on shared disk space.

What will happen now, in theory, is, that when a crystal fails to generate
10000 times in a row:

1. The error message will have printed 10000 times to the log file.
2. The program will print a message which indicates to us where exactly
   in the code the issue is occurring.
3. The program will crash.

That's my guess as to what will happen. We will see what actually happens
next.

It seems this bug occurs mostly in CLI mode.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>